test sorting updated
